Add files via upload
I was doing the database table_Id auto generation function, when I found an error in the 
String findId_sql = "SELECT MAX(" + ID_type +") FROM " + table_name;
Ids.next();
Id = Ids.getInt(1);
Id ++;

It is supposed to provide me with the MaxId+1; In one case it does run out the result as the expectation, but in another case it does not. 
I cannot figure out this bug now.